Enable shell completion of `ucm`
The `postInstall` was never getting executed (because `installPhase` was missing
`runHook postInstall`). So I added the missing `runHooks`.

However, that errored, as the `--runtime-path` (set by the wrapper) and
`--*-completion-script` options don’t work together. To avoid that, I generated
the completions for the wrapped command from the unwrapped command.