Extract the Home Manager config
It was getting too long to keep in the flake, and it works better as an example
on its own.

I also documented the HM config a bit, and made some changes that made more
sense once it was extracted:

- added a tree-sitter grammar derivation to packages;
- moved `self.lib.tree-sitter-grammars` to `self.overlays.tree-sitter`;
- overlaid the Unison tree-sitter grammar on the `nvim-treesitter` Vim plugin;
- added our overlay to Nixpkgs in the HM config, rather than `import nixpkgs …`;
- split up the Home Manager sections to make the distinct configs more obvious;
- made the Emacs’ config _slightly_ less annoying;
- removed `programs.emacs.package` (Emacs 29 is the default now);
- added `programs.neovim` to the HM config; and
- updated `home.stateVersion`.

[^1]: I was hoping to get rid of the former, since there is now a derivation for
the grammar, but it seems `pkgs.tree-sitter` wants to make all its derivations
itself.