Disable shell completion for now
It doesn't work on nixpkgs-unstable, and I've failed to figure out why.
It claims that the `unison` binary isn't found even though it's there,
which I think actually means "this is Nix and you are running prebuilt
binary and I'm angry." I thought that maybe moving the shell completion
to the `postFixup` would help since it should run after elf patching,
but that didn't seem to help at all.