Add the source derivation for Unison
This renames the existng `ucm` outputs to `ucm-bin` and adds a new `ucm` output
for the built-from-source version (pinned to the latest release, 0.5.36). The
new substituter info should also make it so that users don’t have to actually
build the source derivation.

This also moves all references to the `ucm` derivation to `ucm-bin` (e.g., the
overlay has `unison-ucm = localPkgs.ucm-bin`) in order to avoid any issues
introduced by differences between the two packages.